// packages/cross-chain/src/lib/generated/v1/tx.ts
export enum vmType {
  EVM = 0,
  SVM = 1,
  MOVE_VM = 2,
  WASM_VM = 3,
  CAIRO_VM = 4,
  OTHER_VM = 5,
  UNRECOGNIZED = -1,
}
// ...
export interface AccountId {
  namespace: string;
  chainId: string;
  ownerKey: string;
  vmType: vmType;
}

export interface MsgDeployNMSC {
  signer: string;
  accountId: AccountId | undefined;
  txHash: string;
}

export interface MsgMintPush {
  signer: string;
  /** account_id is the identifier of the crosschain owner account */
  accountId:
    | AccountId
    | undefined;
  /** tx_hash is the hash of the transaction in which user locked the tokens */
  txHash: string;
}

export interface CrossChainPayload {
  /** EVM address as hex string (0x...) */
  target: string;
  /** Amount in wei as string (uint256) */
  value: string;
  /** ABI-encoded calldata */
  data: string;
  /** uint256 as string */
  gasLimit: string;
  /** uint256 as string */
  maxFeePerGas: string;
  /** uint256 as string */
  maxPriorityFeePerGas: string;
  /** uint256 as string */
  nonce: string;
  /** uint256 as string */
  deadline: string;
}

export interface MsgExecutePayload {
  signer: string;
  /** account_id is the identifier of the crosschain owner account */
  accountId:
    | AccountId
    | undefined;
  /** payload is the crosschain payload to be executed */
  crosschainPayload:
    | CrossChainPayload
    | undefined;
  /** signature is the signature of the payload by user */
  signature: string;
}
// ...
export const MsgDeployNMSC: MessageFns<MsgDeployNMSC> = {
  encode(message: MsgDeployNMSC, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signer !== "") {
      writer.uint32(10).string(message.signer);
    }
    if (message.accountId !== undefined) {
      AccountId.encode(message.accountId, writer.uint32(18).fork()).join();
    }
    if (message.txHash !== "") {
      writer.uint32(26).string(message.txHash);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgDeployNMSC {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgDeployNMSC();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signer = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.accountId = AccountId.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.txHash = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgDeployNMSC {
    return {
      signer: isSet(object.signer) ? globalThis.String(object.signer) : "",
      accountId: isSet(object.accountId) ? AccountId.fromJSON(object.accountId) : undefined,
      txHash: isSet(object.txHash) ? globalThis.String(object.txHash) : "",
    };
  },

  toJSON(message: MsgDeployNMSC): unknown {
    const obj: any = {};
    if (message.signer !== "") {
      obj.signer = message.signer;
    }
    if (message.accountId !== undefined) {
      obj.accountId = AccountId.toJSON(message.accountId);
    }
    if (message.txHash !== "") {
      obj.txHash = message.txHash;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgDeployNMSC>, I>>(base?: I): MsgDeployNMSC {
    return MsgDeployNMSC.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgDeployNMSC>, I>>(object: I): MsgDeployNMSC {
    const message = createBaseMsgDeployNMSC();
    message.signer = object.signer ?? "";
    message.accountId = (object.accountId !== undefined && object.accountId !== null)
      ? AccountId.fromPartial(object.accountId)
      : undefined;
    message.txHash = object.txHash ?? "";
    return message;
  },
};

function createBaseMsgMintPush(): MsgMintPush {
  return { signer: "", accountId: undefined, txHash: "" };
}

export const MsgMintPush: MessageFns<MsgMintPush> = {
  encode(message: MsgMintPush, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signer !== "") {
      writer.uint32(10).string(message.signer);
    }
    if (message.accountId !== undefined) {
      AccountId.encode(message.accountId, writer.uint32(18).fork()).join();
    }
    if (message.txHash !== "") {
      writer.uint32(26).string(message.txHash);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgMintPush {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgMintPush();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signer = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.accountId = AccountId.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.txHash = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgMintPush {
    return {
      signer: isSet(object.signer) ? globalThis.String(object.signer) : "",
      accountId: isSet(object.accountId) ? AccountId.fromJSON(object.accountId) : undefined,
      txHash: isSet(object.txHash) ? globalThis.String(object.txHash) : "",
    };
  },

  toJSON(message: MsgMintPush): unknown {
    const obj: any = {};
    if (message.signer !== "") {
      obj.signer = message.signer;
    }
    if (message.accountId !== undefined) {
      obj.accountId = AccountId.toJSON(message.accountId);
    }
    if (message.txHash !== "") {
      obj.txHash = message.txHash;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgMintPush>, I>>(base?: I): MsgMintPush {
    return MsgMintPush.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgMintPush>, I>>(object: I): MsgMintPush {
    const message = createBaseMsgMintPush();
    message.signer = object.signer ?? "";
    message.accountId = (object.accountId !== undefined && object.accountId !== null)
      ? AccountId.fromPartial(object.accountId)
      : undefined;
    message.txHash = object.txHash ?? "";
    return message;
  },
};
// ...
export const MsgExecutePayload: MessageFns<MsgExecutePayload> = {
  encode(message: MsgExecutePayload, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signer !== "") {
      writer.uint32(10).string(message.signer);
    }
    if (message.accountId !== undefined) {
      AccountId.encode(message.accountId, writer.uint32(18).fork()).join();
    }
    if (message.crosschainPayload !== undefined) {
      CrossChainPayload.encode(message.crosschainPayload, writer.uint32(26).fork()).join();
    }
    if (message.signature !== "") {
      writer.uint32(34).string(message.signature);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgExecutePayload {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgExecutePayload();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signer = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.accountId = AccountId.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.crosschainPayload = CrossChainPayload.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.signature = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgExecutePayload {
    return {
      signer: isSet(object.signer) ? globalThis.String(object.signer) : "",
      accountId: isSet(object.accountId) ? AccountId.fromJSON(object.accountId) : undefined,
      crosschainPayload: isSet(object.crosschainPayload)
        ? CrossChainPayload.fromJSON(object.crosschainPayload)
        : undefined,
      signature: isSet(object.signature) ? globalThis.String(object.signature) : "",
    };
  },

  toJSON(message: MsgExecutePayload): unknown {
    const obj: any = {};
    if (message.signer !== "") {
      obj.signer = message.signer;
    }
    if (message.accountId !== undefined) {
      obj.accountId = AccountId.toJSON(message.accountId);
    }
    if (message.crosschainPayload !== undefined) {
      obj.crosschainPayload = CrossChainPayload.toJSON(message.crosschainPayload);
    }
    if (message.signature !== "") {
      obj.signature = message.signature;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgExecutePayload>, I>>(base?: I): MsgExecutePayload {
    return MsgExecutePayload.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgExecutePayload>, I>>(object: I): MsgExecutePayload {
    const message = createBaseMsgExecutePayload();
    message.signer = object.signer ?? "";
    message.accountId = (object.accountId !== undefined && object.accountId !== null)
      ? AccountId.fromPartial(object.accountId)
      : undefined;
    message.crosschainPayload = (object.crosschainPayload !== undefined && object.crosschainPayload !== null)
      ? CrossChainPayload.fromPartial(object.crosschainPayload)
      : undefined;
    message.signature = object.signature ?? "";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}